// be/src/domain/entities/Category.d.ts
export declare class Category {
    private id;
    private name;
    private icon;
    private createdAt;
    constructor(id: string, name: string, icon: string, createdAt?: string);
    getId(): string;
    getName(): string;
    getIcon(): string;
    getCreatedAt(): string;
}
//# sourceMappingURL=Category.d.ts.map